Tweaking products route - it still doesn't work
