+++ client/src/pages/Home.jsx
@@ -1,32 +1,10 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import ProductList from '../components/product/ProductList/ProductList';
-import axios from 'axios';
+import useProducts from '../hooks/useProducts';
 import { Card } from '@mui/material';
 
-const API_URL = process.env.REACT_APP_API_URL || 'http://localhost:5000/api';
-
-
 const Home = () => {
-	const [products, setProducts] = useState([]);
-	const [loading, setLoading] = useState(true);
-	const [error, setError] = useState(null);
-
-	useEffect(() => {
-		const fetchProducts = async () => {
-			try {
-				const response = await axios.get(`${API_URL}/products`);
-				setProducts(response.data);
-				setLoading(false);
-			} catch (err) {
-				console.error('Error fetching products:', err);
-				setError('Failed to load products. Please try again later.');
-				setLoading(false);
-			}
-		};
-
-		fetchProducts();
-	}, []);
-
+	const { products, loading, error } = useProducts();
 	if (loading) {
 		return <div>Loading products...</div>;
 	}
