+++ server/models/Product.js
@@ -2,55 +2,55 @@ const { DataTypes } = require('sequelize');
 const { sequelize } = require('../config/database');
 
 const Product = sequelize.define('Product', {
-    id: {
-        type: DataTypes.INTEGER,
-        primaryKey: true,
-        autoIncrement: true
-    },
-    name: {
-        type: DataTypes.STRING,
-        allowNull: false,
-        validate: {
-            notEmpty: true
-        }
-    },
-    description: {
-        type: DataTypes.TEXT,
-        allowNull: false,
-        validate: {
-            notEmpty: true
-        }
-    },
-    price: {
-        type: DataTypes.DECIMAL(10, 2),
-        allowNull: false,
-        validate: {
-            min: 0
-        }
-    },
-    imageUrl: {
-        type: DataTypes.STRING,
-        allowNull: false,
-        validate: {
-            notEmpty: true
-        }
-    },
-    category: {
-        type: DataTypes.STRING,
-        allowNull: false,
-        validate: {
-            notEmpty: true
-        }
-    },
-    stock: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        validate: {
-            min: 0
-        }
-    }
+	id: {
+		type: DataTypes.INTEGER,
+		primaryKey: true,
+		autoIncrement: true
+	},
+	name: {
+		type: DataTypes.STRING,
+		allowNull: false,
+		validate: {
+			notEmpty: true
+		}
+	},
+	description: {
+		type: DataTypes.TEXT,
+		allowNull: false,
+		validate: {
+			notEmpty: true
+		}
+	},
+	price: {
+		type: DataTypes.DECIMAL(10, 2),
+		allowNull: false,
+		validate: {
+			min: 0
+		}
+	},
+	imageUrl: {
+		type: DataTypes.STRING,
+		allowNull: false,
+		validate: {
+			notEmpty: true
+		}
+	},
+	category: {
+		type: DataTypes.STRING,
+		allowNull: false,
+		validate: {
+			notEmpty: true
+		}
+	},
+	stock: {
+		type: DataTypes.INTEGER,
+		allowNull: false,
+		validate: {
+			min: 0
+		}
+	}
 }, {
-    timestamps: true // This will add createdAt and updatedAt automatically
+	timestamps: true // This will add createdAt and updatedAt automatically
 });
 
 module.exports = Product;
